requirements.txt
django>=4.2
djangorestframework
django-environ
redis
uvicorn
psycopg2-binary
pytest
pytest-django
httpx
grpcio
grpcio-tools
grpcio-testing
